Remove context subcrate
* Add components module
* Add world_resources module
* Remove gfx_device::assets module
