+++ Cargo.toml
@@ -16,10 +16,6 @@ license = "MIT OR Apache-2.0"
 path = "src/config/"
 version = "0.1.0"
 
-[dependencies.amethyst_context]
-path = "src/context/"
-version = "0.1.0"
-
 [dependencies.amethyst_ecs]
 path = "src/ecs/"
 version = "0.1.1"
