+++ examples/02_sphere/main.rs
@@ -9,11 +9,8 @@ use amethyst::config::Element;
 use amethyst::ecs::World;
 use amethyst::gfx_device::DisplayConfig;
 use amethyst::asset_manager::AssetManager;
-use amethyst::context::event::EngineEvent;
-use amethyst::gfx_device::assets::{Texture, Mesh};
-use amethyst::gfx_device::Renderable;
+use amethyst::components::event::EngineEvent;
 use amethyst::renderer::{VertexPosNormal, Pipeline};
-use amethyst::processors::transform::LocalTransform;
 
 use self::genmesh::generators::{SphereUV};
 use self::genmesh::{MapToVertices, Triangulate, Vertices};
@@ -23,10 +20,12 @@ struct Example;
 
 impl State for Example {
     fn on_start(&mut self, world: &mut World, asset_manager: &mut AssetManager, pipeline: &mut Pipeline) {
-        use amethyst::renderer::pass::*;
+        use amethyst::renderer::pass::{Clear, DrawShaded};
         use amethyst::renderer::{Layer, Light};
-        use amethyst::gfx_device::world_resources::camera::*;
-        use amethyst::gfx_device::world_resources::ScreenDimensions;
+        use amethyst::world_resources::camera::{Projection, Camera};
+        use amethyst::world_resources::ScreenDimensions;
+        use amethyst::components::transform::LocalTransform;
+        use amethyst::components::rendering::{Texture, Mesh, Renderable};
         let clear_layer =
             Layer::new("main",
                         vec![
@@ -74,7 +73,7 @@ impl State for Example {
     }
 
     fn handle_events(&mut self, events: &[EngineEvent], _: &mut World, _: &mut AssetManager, _: &mut Pipeline) -> Trans {
-        use amethyst::context::event::*;
+        use amethyst::components::event::*;
         for event in events {
             match event.payload {
                 Event::KeyboardInput(_, _, Some(VirtualKeyCode::Escape)) => return Trans::Quit,
