+++ src/components/event/event.rs
@@ -7,8 +7,7 @@ use self::amethyst_ecs::{Component, VecStorage};
 pub use self::glutin::{Event, ElementState, ScanCode, VirtualKeyCode, MouseScrollDelta, TouchPhase, MouseButton, Touch};
 
 /// Represents an engine generated event,
-/// it can be attached to entities
-/// that are published by `Broadcaster`.
+/// it can be attached to published by `Broadcaster`.
 /// Currently it is just a wraper around
 /// `glutin::Event`.
 pub struct EngineEvent {
