+++ src/components/event/mod.rs
@@ -1,0 +1,3 @@
+mod event;
+
+pub use self::event::*;
