+++ src/components/rendering/mod.rs
@@ -1,0 +1,7 @@
+mod texture;
+mod mesh;
+mod renderable;
+
+pub use self::texture::*;
+pub use self::mesh::*;
+pub use self::renderable::Renderable;
