+++ src/components/transform/transform.rs
@@ -1,0 +1,31 @@
+use ecs::{Component, VecStorage};
+
+/// Absolute transformation (transformed from origin).
+/// Used for rendering position and orientation.
+#[derive(Debug, Copy, Clone)]
+pub struct Transform(pub [[f32; 4]; 4]);
+
+impl Component for Transform {
+    type Storage = VecStorage<Transform>;
+}
+
+impl Default for Transform {
+    fn default() -> Self {
+        Transform([[1.0, 0.0, 0.0, 0.0],
+                   [0.0, 1.0, 0.0, 0.0],
+                   [0.0, 0.0, 1.0, 0.0],
+                   [0.0, 0.0, 0.0, 1.0]])
+    }
+}
+
+impl From<[[f32; 4]; 4]> for Transform {
+    fn from(matrix: [[f32; 4]; 4]) -> Self {
+        Transform(matrix)
+    }
+}
+
+impl Into<[[f32; 4]; 4]> for Transform {
+    fn into(self) -> [[f32; 4]; 4] {
+        self.0
+    }
+}
