--- src/context/Cargo.toml
@@ -1,34 +1,0 @@
-[package]
-name = "amethyst_context"
-version = "0.1.0"
-authors = ["Nikita Chashchinskii <nikita.chashinsky@yandex.ru>"]
-description = "A way to manage engine resources"
-keywords = ["game", "engine", "amethyst"]
-
-documentation = "https://www.amethyst.rs/doc/amethyst/"
-homepage = "https://www.amethyst.rs/"
-repository = "https://github.com/amethyst/amethyst"
-
-license = "MIT OR Apache-2.0"
-
-[dependencies.amethyst_renderer]
-path = "../renderer/"
-version = "0.3.0"
-
-[dependencies.amethyst_config]
-path = "../config/"
-version = "0.1.0"
-
-[dependencies.amethyst_ecs]
-path = "../ecs/"
-version = "0.1.0"
-
-[dependencies]
-gfx = "0.12"
-gfx_core = "0.4"
-gfx_device_gl = "0.11"
-gfx_window_glutin = "0.12"
-glutin = "0.6"
-time = "0.1"
-genmesh="0.4"
-cgmath="0.11"
