--- src/gfx_device/assets/mod.rs
@@ -1,5 +1,0 @@
-pub mod texture;
-pub mod mesh;
-
-pub use self::texture::Texture;
-pub use self::mesh::Mesh;
