+++ src/gfx_device/mod.rs
@@ -8,14 +8,10 @@ mod gfx_device;
 mod main_target;
 mod video_init;
 mod display_config;
-mod renderable;
 
-pub mod assets;
 pub mod gfx_loader;
-pub mod world_resources;
 
 pub use self::display_config::DisplayConfig;
 pub use self::video_init::video_init;
 pub use self::main_target::*;
 pub use self::gfx_device::*;
-pub use self::renderable::Renderable;
