+++ src/processors/transform.rs
@@ -3,183 +3,11 @@
 extern crate cgmath;
 // extern crate test;
 
-use self::cgmath::{Quaternion, Vector3, Matrix3, Matrix4};
+use self::cgmath::Matrix4;
 
-use ecs::{Join, Component, NullStorage, VecStorage, Entity, RunArg, Processor};
-// use context::Context;
-use std::sync::atomic::{AtomicBool, Ordering};
+use ecs::{Join, Entity, RunArg, Processor};
+use components::transform::{LocalTransform, Transform, Child, Init};
 use std::collections::{HashMap, HashSet};
-use std::ops::{Deref, DerefMut};
-
-#[derive(Debug)]
-pub struct InnerTransform {
-    /// Translation/position vector [x, y, z]
-    pub translation: [f32; 3],
-
-    /// Quaternion [w (scalar), x, y, z]
-    pub rotation: [f32; 4],
-
-    /// Scale vector [x, y, z]
-    pub scale: [f32; 3],
-}
-
-/// Local position, rotation, and scale (from parent if it exists).
-#[derive(Debug)]
-pub struct LocalTransform {
-    /// Wrapper around the transform data for dirty flag setting.
-    wrapped: InnerTransform,
-
-    /// Flag for re-computation
-    dirty: AtomicBool,
-}
-
-impl Deref for LocalTransform {
-    type Target = InnerTransform;
-    fn deref(&self) -> &InnerTransform {
-        &self.wrapped
-    }
-}
-
-impl DerefMut for LocalTransform {
-    fn deref_mut(&mut self) -> &mut InnerTransform {
-        self.flag(true);
-        &mut self.wrapped
-    }
-}
-
-impl LocalTransform {
-    /// Flags the current transform for re-computation.
-    ///
-    /// Note: All `set_*` methods will automatically flag the component.
-    #[inline]
-    pub fn flag(&self, dirty: bool) {
-        self.dirty.store(dirty, Ordering::SeqCst);
-    }
-
-    /// Returns whether or not the current transform is flagged for re-computation or "dirty".
-    #[inline]
-    pub fn is_dirty(&self) -> bool {
-        self.dirty.load(Ordering::SeqCst)
-    }
-
-    /// Returns the local object matrix for the transform.
-    ///
-    /// Combined with the parent's global `Transform` component it gives
-    /// the global (or world) matrix for the current entity.
-    #[inline]
-    pub fn matrix(&self) -> [[f32; 4]; 4] {
-        let quat: Matrix3<f32> = Quaternion::from(self.rotation).into();
-        let scale: Matrix3<f32> = Matrix3::<f32> {
-            x: [self.scale[0], 0.0, 0.0].into(),
-            y: [0.0, self.scale[1], 0.0].into(),
-            z: [0.0, 0.0, self.scale[2]].into(),
-        };
-        let mut matrix: Matrix4<f32> = (&quat * scale).into();
-        matrix.w = Vector3::from(self.translation).extend(1.0f32);
-        matrix.into()
-    }
-}
-
-impl Default for LocalTransform {
-    fn default() -> Self {
-        LocalTransform {
-            wrapped: InnerTransform {
-                translation: [0.0, 0.0, 0.0],
-                rotation: [1.0, 0.0, 0.0, 0.0],
-                scale: [1.0, 1.0, 1.0],
-            },
-            dirty: AtomicBool::new(true),
-        }
-    }
-}
-
-impl Component for LocalTransform {
-    type Storage = VecStorage<LocalTransform>;
-}
-
-/// Absolute transformation (transformed from origin).
-/// Used for rendering position and orientation.
-#[derive(Debug, Copy, Clone)]
-pub struct Transform(pub [[f32; 4]; 4]);
-
-impl Component for Transform {
-    type Storage = VecStorage<Transform>;
-}
-
-impl Default for Transform {
-    fn default() -> Self {
-        Transform([[1.0, 0.0, 0.0, 0.0],
-                   [0.0, 1.0, 0.0, 0.0],
-                   [0.0, 0.0, 1.0, 0.0],
-                   [0.0, 0.0, 0.0, 1.0]])
-    }
-}
-
-impl From<[[f32; 4]; 4]> for Transform {
-    fn from(matrix: [[f32; 4]; 4]) -> Self {
-        Transform(matrix)
-    }
-}
-
-impl Into<[[f32; 4]; 4]> for Transform {
-    fn into(self) -> [[f32; 4]; 4] {
-        self.0
-    }
-}
-
-/// Initialization flag.
-/// Added to entity with a `LocalTransform` component after the first update.
-#[derive(Default, Copy, Clone)]
-pub struct Init;
-impl Component for Init {
-    type Storage = NullStorage<Init>;
-}
-
-/// Component for defining a parent entity.
-pub struct Child {
-    /// The parent entity
-    parent: Entity,
-
-    /// Flag for whether the child was changed
-    dirty: AtomicBool,
-}
-
-impl Child {
-    pub fn new(entity: Entity) -> Child {
-        Child {
-            parent: entity,
-            dirty: AtomicBool::new(true),
-        }
-    }
-
-    #[inline]
-    pub fn parent(&self) -> Entity {
-        self.parent
-    }
-    #[inline]
-    pub fn set_parent(&mut self, entity: Entity) {
-        self.parent = entity;
-        self.flag(true);
-    }
-
-    /// Flag that parent has been changed
-    ///
-    /// Note: `set_parent` flags the parent.
-    #[inline]
-    pub fn flag(&self, dirty: bool) {
-        self.dirty.store(dirty, Ordering::SeqCst);
-    }
-
-    /// Returns whether the parent was changed.
-    #[inline]
-    pub fn is_dirty(&self) -> bool {
-        self.dirty.load(Ordering::SeqCst)
-    }
-}
-
-impl Component for Child {
-    type Storage = VecStorage<Child>;
-}
 
 /// Transformation processor.
 ///
